back/flask_server/requirements.txt
torch
torchvision
matplotlib
tqdm
pytesseract
Pillow
numpy
tokenizers
pandas
opencv-python
requests
python-socketio
Flask
Flask-CORS
Flask-SocketIO
